Add data to awards vocabulary
