+++ app_data/vocabularies/funders.yaml
@@ -1,0 +1,184 @@
+- country: GB
+  id: 0439y7842
+  identifiers:
+  - identifier: https://ror.org/0439y7842
+    scheme: ror
+  name: Engineering and Physical Sciences Research Council
+  title:
+    en: Engineering and Physical Sciences Research Council
+- country: GB
+  id: 03wnrjx87
+  identifiers:
+  - identifier: https://ror.org/03wnrjx87
+    scheme: ror
+  name: Royal Society
+  title:
+    en: Royal Society
+- country: GB
+  id: 03x94j517
+  identifiers:
+  - identifier: https://ror.org/03x94j517
+    scheme: ror
+  name: Medical Research Council
+  title:
+    en: Medical Research Council
+- country: GB
+  id: 056ffv270
+  identifiers:
+  - identifier: https://ror.org/056ffv270
+    scheme: ror
+  name: Imperial College Healthcare NHS Trust
+  title:
+    en: Imperial College Healthcare NHS Trust
+- country: GB
+  id: 0187kwz08
+  identifiers:
+  - identifier: https://ror.org/0187kwz08
+    scheme: ror
+  name: National Institute for Health Research
+  title:
+    en: National Institute for Health Research
+- country: GB
+  id: 029chgv08
+  identifiers:
+  - identifier: https://ror.org/029chgv08
+    scheme: ror
+  name: Wellcome Trust
+  title:
+    en: Wellcome Trust
+- country: BE
+  id: 00k4n6c32
+  identifiers:
+  - identifier: https://ror.org/00k4n6c32
+    scheme: ror
+  name: European Commission
+  title:
+    en: European Commission
+- country: GB
+  id: 02wdwnk04
+  identifiers:
+  - identifier: https://ror.org/02wdwnk04
+    scheme: ror
+  name: British Heart Foundation
+  title:
+    en: British Heart Foundation
+- country: GB
+  id: 054225q67
+  identifiers:
+  - identifier: https://ror.org/054225q67
+    scheme: ror
+  name: Cancer Research UK
+  title:
+    en: Cancer Research UK
+- country: GB
+  id: 00cwqg982
+  identifiers:
+  - identifier: https://ror.org/00cwqg982
+    scheme: ror
+  name: Biotechnology and Biological Sciences Research Council
+  title:
+    en: Biotechnology and Biological Sciences Research Council
+- country: GB
+  id: 057g20z61
+  identifiers:
+  - identifier: https://ror.org/057g20z61
+    scheme: ror
+  name: Science and Technology Facilities Council
+  title:
+    en: Science and Technology Facilities Council
+- country: GB
+  id: 05ar5fy68
+  identifiers:
+  - identifier: https://ror.org/05ar5fy68
+    scheme: ror
+  name: Innovate UK
+  title:
+    en: Innovate UK
+- country: GB
+  id: 02b5d8509
+  identifiers:
+  - identifier: https://ror.org/02b5d8509
+    scheme: ror
+  name: Natural Environment Research Council
+  title:
+    en: Natural Environment Research Council
+- country: GB
+  id: 04e3zg361
+  identifiers:
+  - identifier: https://ror.org/04e3zg361
+    scheme: ror
+  name: Rosetrees Trust
+  title:
+    en: Rosetrees Trust
+- country: GB
+  id: 001aqnf71
+  identifiers:
+  - identifier: https://ror.org/001aqnf71
+    scheme: ror
+  name: UK Research and Innovation
+  title:
+    en: UK Research and Innovation
+- country: GB
+  id: 012mzw131
+  identifiers:
+  - identifier: https://ror.org/012mzw131
+    scheme: ror
+  name: Leverhulme Trust
+  title:
+    en: Leverhulme Trust
+- country: US
+  id: 01cwqze88
+  identifiers:
+  - identifier: https://ror.org/01cwqze88
+    scheme: ror
+  name: National Institutes of Health
+  title:
+    en: National Institutes of Health
+- country: GB
+  id: 0439y7842
+  identifiers:
+  - identifier: https://ror.org/0439y7842
+    scheme: ror
+  name: Engineering and Physical Sciences Research Council
+  title:
+    en: Engineering and Physical Sciences Research Council
+- country: US
+  id: 0456r8d26
+  identifiers:
+  - identifier: https://ror.org/0456r8d26
+    scheme: ror
+  name: Gates Foundation
+  title:
+    en: Gates Foundation
+- country: GB
+  id: 02wedp412
+  identifiers:
+  - identifier: https://ror.org/02wedp412
+    scheme: ror
+  name: UK Dementia Research Institute
+  title:
+    en: UK Dementia Research Institute
+- country: GB
+  id: 04jswqb94
+  identifiers:
+  - identifier: https://ror.org/04jswqb94
+    scheme: ror
+  name: Defence Science and Technology Laboratory
+  title:
+    en: Defence Science and Technology Laboratory
+- country: GB
+  id: 00c489v88
+  identifiers:
+  - identifier: https://ror.org/00c489v88
+    scheme: ror
+  name: Academy of Medical Sciences
+  title:
+    en: Academy of Medical Sciences
+- country: GB
+  id: 05dt4bt98
+  identifiers:
+  - identifier: https://ror.org/05dt4bt98
+    scheme: ror
+  name: The Faraday Institution
+  title:
+    en: The Faraday Institution
